requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.4
h11==0.16.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.37.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.18
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
requests>=2.25.1
backoff>=1.10.0
python-json-logger>=2.0.7
sentry-sdk[fastapi]==2.35.0
upstash-redis==1.4.0
upstash-ratelimit==1.1.0